+++ constants.js
@@ -3,10 +3,12 @@ const NOTION_URL = 'https://www.notion.so';
 
 const TAKE_SCREEN_SHOT = 'TAKE_SCREEN_SHOT';
 const GET_VIDEO_POSITION_DATA = 'GET_VIDEO_POSITION_DATA';
-const TOGGLE_CAPTURE_AREA = 'TOGGLE_CAPTURE_AREA';
+const DISABLE_CAPTURE_AREA = 'DISABLE_CAPTURE_AREA';
+const TOGGLE_SHOW_HIDE_CAPTURE_AREA = 'TOGGLE_SHOW_HIDE_CAPTURE_AREA';
 // Get screen width (screen width should only be obtained from content script not popup)
 const GET_SCREEN_DIMENSIONS = 'GET_SCREEN_DIMENSIONS';
 const SHOW_CAPTURE_AREA_IF_NO_VIDEO = 'SHOW_CAPTURE_AREA_IF_NO_VIDEO';
+const PLAY_PAUSE_VIDEO = 'PLAY_PAUSE_VIDEO';
 
 const NOTION_ERR_MESSAGE = `Error 😢. Try 
   1️⃣ Refresh the video window
@@ -18,6 +20,7 @@ const FEEDBACK_URL = 'https://docs.google.com/forms/d/e/1FAIpQLSdq6KtGZKvPLN2KBa
 const REVIEW_URL = 'https://chrome.google.com/webstore/detail/sidenotion/ihjmeedcmbeapmaagcjpngemnckiooih/reviews';
 
 const USAGE_HINT_CONTAINER_ID = 'usage-hint-container';
+const PLAY_PAUSE_HINT_ID = 'play-pause-hint';
 const CAPTURE_HINT_ID = 'capture-hint';
-const TUTORIAL_HINT_ID = 'tutorial-hint';
-const TOGGLE_CAPTURE_AREA_HINT_ID = 'toggle-capture-area-hint';
+const SHOW_TUTORIAL_HINT_ID = 'tutorial-hint';
+const SHOW_HIDE_AREA_HINT_ID = 'toggle-capture-area-hint';